+++ CHANGELOG.md
@@ -6,9 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.0.0] - 2021-01-09
 ### Changed
 #### Scrobble timeline
 * [breaking] Dataset format update: fetched scrobbles list is now an array of arrays (`[[date, trackName, trackPlaycount, albumName, albumPlaycount, artistName, artistPlaycount], ...]`), not an array of objects.
+* Fixed: timezone offsets were applied incorrectly.
 
 ## [1.3.0] - 2020-01-19
 ### Added